requirements.txt
litserve==0.2.2
qdrant-client==1.11.2
python-dotenv==1.0.1
arize-phoenix==4.35.1
llama-index==0.11.10
llama-index-llms-openai==0.2.8
llama-index-embeddings-fastembed==0.2.0
llama-index-vector-stores-qdrant==0.3.0
llama-index-callbacks-arize-phoenix==0.2.1
llama-index-readers-azstorage-blob==0.2.0
pydantic==2.9.0
uvloop==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
